requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
redis==5.0.4
psycopg2-binary==2.9.9
pydantic-settings==2.2.1
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
